+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to PyToolsmith will be documented in this file.
 
+## 0.1.3 - March 9, 2025
+
+### Added
+
+- Basic tool calling support, without custom serialization.
+
+## 0.1.2 - March 9, 2025
+
+### Added
+
+- Dataclasses to define most major types
+- Additional tests for tool library
+
+### Updated
+
+- Removed dependency on pydantic to make library as lightweight as possible
+
 ## 0.1.1 - March 7, 2025
 
 ### Added